Interaction with MiniPlayer and component filter item
