+++ my_music_code/lib/Search/Components/query_result_widget.dart
