+++ my_music_code/lib/Search/search_page_terms.dart
@@ -44,17 +44,14 @@ class _SearchPageTermsState extends State<SearchPageTerms> {
     setState(() {
       buildTile = List.from(listResults.map((music) => ListTile(
           onTap: () {
-            setState(() {
-              
-            });
             showModalBottomSheet(
-                useRootNavigator: false,
-                isScrollControlled: true,
-                useSafeArea: true,
-                context: context,
-                builder: (context) {
-                  return MusicPage(music: music);
-                });
+              useRootNavigator: false,
+              isScrollControlled: true,
+              useSafeArea: true,
+              context: context,
+              builder: (context) {
+                return MusicPage(music: music);
+              });
           },
           title: Text(music.name!, style: TextStyle(color: Colors.white, fontWeight: FontWeight.w500)),
           subtitle: Text(music.artist!, style: TextStyle(color: Colors.white)),
